# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte ETV",
        "description": "REST API. Projecte ETV. DAW Client i servidor.",
        "contact": {
            "name": "Webmaster.",
            "email": "info@webmaster.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.dawpaucasesnoves.com/etvServidor/public"
        }
    ],
    "paths": {
...
        "/api/audios": {
            "get": {
                "tags": [
                    "Audios"
                ],
                "summary": "Mostrar els audios",
                "operationId": "6438c4417715b0d99d1f111214fb5fdc",
                "responses": {
                    "200": {
                        "description": "Mostrar tots els audios.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Àudios"
                ],
                "summary": "Crea un àudio",
                "description": "Crea un nou audio a la base de dades i el puja al servidor.",
                "operationId": "824214a2f963cba05d48ebca6fae29a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ordre": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "transcripcio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Transcripció d'un àudio"
                                    },
                                    "fk_puntInteres": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/audios/{id}": {
            "get": {
                "tags": [
                    "Audios"
                ],
                "summary": "Mostrar un audio",
                "operationId": "a3719e15994b078d89eb799ccbf3ccdf",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "missatge": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string"
                                        },
                                        "codi": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Audios"
                ],
                "summary": "Modifica un audio",
                "description": "Modifica un audio. Sols per administradors o el gestor del punt d'interes",
                "operationId": "38b6255bf40be6af0c893f9999c443e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Àudio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ordre": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "transcripcio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Transcripció d'un àudio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "missatge": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string"
                                        },
                                        "codi": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Àudios"
                ],
                "summary": "Esborra un àudio",
                "description": "Esborra un àudio. Sols per administradors o gestors autors de l'audio",
                "operationId": "93f2f39f35540f1b19f92fc178ea6963",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Àudio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "missatge": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        },
                                        "missatge": {
                                            "type": "string",
                                            "example": "Unprocessable Entity"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "missatge": {
                                            "type": "string"
                                        },
                                        "codi": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}